pyproject.toml
[project]
name = "mkdocs-s3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib==2.201.0",
    "constructs>=10.0.0,<11.0.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest==6.2.5",
]
